chore: private 지정 및 extension 정리, delegate 분리
